pyproject.toml
[project]
name = "memex-mcp"
version = "0.1.0"
description = "MemexMCP: local semantic + FTS search over your codebases and documents, served over MCP."
requires-python = ">=3.12"
dependencies = [
    "sqlite-vec>=0.1.9",
    "httpx>=0.27",
    "typer>=0.12",
    "mcp>=1.2",
    "keyring>=25",
    "pypdf>=5",
    "python-docx>=1.1",
    "selectolax>=0.3",
]

[project.scripts]
memex-cli = "memex.cli:app"
memex-mcp-server = "memex.mcp_server:main"

[project.gui-scripts]
Memex = "memex.gui:main"

[project.optional-dependencies]
gui = ["PySide6>=6.7"]
dev = ["ruff>=0.6", "pytest>=8", "pyinstaller>=6.10"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/memex"]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.ruff.lint.flake8-bugbear]
extend-immutable-calls = ["typer.Argument", "typer.Option"]
